requirements.txt
pandas==1.3.4
beautifulsoup4==4.10.0
cryptography==36.0.0
Office365-REST-Python-Client==2.3.9
openpyxl==3.0.9
ipykernel==6.6.0
pyodbc==4.0.32
html5lib==1.1
lxml==4.7.1
gspread==5.1.1
pyxlsb==1.0.9
xlrd>=1.0.0
selenium==4.8.0
webdriver-manager==3.8.5
matplotlib==3.7.1
matplotlib-inline==0.1.6